[fix] Refactoring
