+++ src/EasyEPlanner.csproj
@@ -174,6 +174,7 @@
       <DependentUpon>ModeFrm.cs</DependentUpon>
     </Compile>
     <Compile Include="PInvokeUtil.cs" />
+    <Compile Include="PrgLuaSaver.cs" />
     <Compile Include="ProjectConfiguration.cs" />
     <Compile Include="ProjectManager.cs" />
     <Compile Include="main.cs" />
