+++ src/ProjectManager.cs
@@ -1384,7 +1384,7 @@ namespace EasyEPlanner
                     "require(\"line\")", 
                     "require(\"master\")"));
                 prgFileWriter.WriteLine("-- Основные объекты проекта (объекты, описанные в Eplan'е).");
-                prgFileWriter.WriteLine(techObjectManager.SavePrgAsLuaTable("\t"));
+                prgFileWriter.WriteLine(PrgLuaSaver.Save("\t"));
 
                 if (par.silentMode == false)
                 {
